Add files via upload
Code is stable and has many structural changes, but largely looks the same. I didn't have a chance to merge the code, so it's here instead of in the main branch for now.
I've been frustrated with my efforts to generate a random map. I think I know how to do it now, but that will be another day.

+++ Thing.cpp
@@ -6,7 +6,7 @@ Thing::Thing(sf::RenderWindow& win) : window(win)
 	x1 = x2 = y1 = y2 = sizeX = sizeY = 0;
 };
 
-bool Thing::isAtLocation(int x, int y)
+bool Thing::isAtLocation(double x, double y)
 {
 	if (((x2 >= x1) && ((x<=x2) && (x>=x1))) || ((x2 < x1) && ((x>=x2) && (x<=x1)))) {	// in range of x
 		if (((y2 >= y1) && ((y <= y2) && (y >= y1))) || ((y2 < y1) && ((y >= y2) && (y <= y1)))) { // in range of y
